// apps/indexer/src/generated/pub/leaflet/pages/linearDocument.defs.ts
type Main = {
  $type?: 'pub.leaflet.pages.linearDocument'
  id?: string
  blocks: Block[]
}
// ...
type Block = {
  $type?: 'pub.leaflet.pages.linearDocument#block'
  block:
    | l.$Typed<BlocksIframe.Main>
    | l.$Typed<BlocksText.Main>
    | l.$Typed<BlocksBlockquote.Main>
    | l.$Typed<BlocksHeader.Main>
    | l.$Typed<BlocksImage.Main>
    | l.$Typed<BlocksUnorderedList.Main>
    | l.$Typed<BlocksWebsite.Main>
    | l.$Typed<BlocksMath.Main>
    | l.$Typed<BlocksCode.Main>
    | l.$Typed<BlocksHorizontalRule.Main>
    | l.$Typed<BlocksBskyPost.Main>
    | l.$Typed<BlocksPage.Main>
    | l.$Typed<BlocksPoll.Main>
    | l.$Typed<BlocksButton.Main>
    | l.Unknown$TypedObject
  alignment?:
    | '#textAlignLeft'
    | '#textAlignCenter'
    | '#textAlignRight'
    | '#textAlignJustify'
    | l.UnknownString
}

export type { Block }

const block = l.typedObject<Block>(
  $nsid,
  'block',
  l.object({
    block: l.typedUnion(
      [
        l.typedRef<BlocksIframe.Main>((() => BlocksIframe.main) as any),
        l.typedRef<BlocksText.Main>((() => BlocksText.main) as any),
        l.typedRef<BlocksBlockquote.Main>((() => BlocksBlockquote.main) as any),
        l.typedRef<BlocksHeader.Main>((() => BlocksHeader.main) as any),
        l.typedRef<BlocksImage.Main>((() => BlocksImage.main) as any),
        l.typedRef<BlocksUnorderedList.Main>(
          (() => BlocksUnorderedList.main) as any,
        ),
        l.typedRef<BlocksWebsite.Main>((() => BlocksWebsite.main) as any),
        l.typedRef<BlocksMath.Main>((() => BlocksMath.main) as any),
        l.typedRef<BlocksCode.Main>((() => BlocksCode.main) as any),
        l.typedRef<BlocksHorizontalRule.Main>(
          (() => BlocksHorizontalRule.main) as any,
        ),
        l.typedRef<BlocksBskyPost.Main>((() => BlocksBskyPost.main) as any),
        l.typedRef<BlocksPage.Main>((() => BlocksPage.main) as any),
        l.typedRef<BlocksPoll.Main>((() => BlocksPoll.main) as any),
        l.typedRef<BlocksButton.Main>((() => BlocksButton.main) as any),
      ],
      false,
    ),
    alignment: l.optional(
      l.string<{
        knownValues: [
          '#textAlignLeft',
          '#textAlignCenter',
          '#textAlignRight',
          '#textAlignJustify',
        ]
      }>(),
    ),
  }),
)

export { block }

type TextAlignLeft = 'pub.leaflet.pages.linearDocument#textAlignLeft'

export type { TextAlignLeft }

const textAlignLeft = l.token($nsid, 'textAlignLeft')

export { textAlignLeft }

type TextAlignCenter = 'pub.leaflet.pages.linearDocument#textAlignCenter'

export type { TextAlignCenter }

const textAlignCenter = l.token($nsid, 'textAlignCenter')

export { textAlignCenter }

type TextAlignRight = 'pub.leaflet.pages.linearDocument#textAlignRight'

export type { TextAlignRight }

const textAlignRight = l.token($nsid, 'textAlignRight')

export { textAlignRight }

type TextAlignJustify = 'pub.leaflet.pages.linearDocument#textAlignJustify'

export type { TextAlignJustify }

const textAlignJustify = l.token($nsid, 'textAlignJustify')